+++ tests/integration/suite/actors/lock/call/local/local.go
@@ -11,8 +11,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package self
+package local
 
 import (
-	_ "github.com/dapr/dapr/tests/integration/suite/actors/lock/call/self/chain"
+	_ "github.com/dapr/dapr/tests/integration/suite/actors/lock/call/local/chain"
 )
